Admin Panel finished

+++ LifeLike.Frontend/src/app/modules/admin/pages/photos/photos.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material';
 import { map } from 'rxjs/operators';
 import Photo from '../../../../modules/photo/models/Photo';
+import { ConfirmDeleteComponent } from '../../dialogs/confirm-delete/confirm-delete.component';
 import { PhotoCreateComponent } from '../../dialogs/photo-create/photo-create.component';
 import { PhotoEditComponent } from '../../dialogs/photo-edit/photo-edit.component';
 import { AdminRestService } from '../../services/admin-rest.service';
@@ -22,17 +23,25 @@ export class PhotosComponent implements OnInit {
   constructor(private readonly restService: AdminRestService, private readonly dialog: MatDialog) { }
 
   Remove(photo: Photo): void {
-    this.IsLoading = true;
-    this.restService.deletePhoto(photo.Id)
-      .subscribe(
-        (data) => {
-          this.IsLoading = false;
-          this.GetPhotos();
-      },
-      (error) => {
-          this.error = error;
-          this.IsLoading = false;
-      });
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.data = photo.Title;
+    dialogConfig.autoFocus = true;
+    const dialogRef = this.dialog.open(ConfirmDeleteComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.IsLoading = true;
+        this.restService.deletePhoto(photo.Id)
+          .subscribe(
+            () => {
+              this.IsLoading = false;
+              this.GetPhotos();
+          },
+          (error) => {
+              this.error = error;
+              this.IsLoading = false;
+          });
+      }
+    });
   }
   Edit(photo: Photo): void {
